+++ pytsite/auth/__init__.py
@@ -29,23 +29,23 @@ def __init():
     # Module permission group
     permissions.define_group('security', 'pytsite.auth@security')
 
-    # Access token HTTP API handlers
-    http_api.handle('POST', 'auth/access-token', _http_api.post_access_token,
+    # Access token HTTP API
+    http_api.handle('POST', 'auth/access-token/<driver>', _http_api.post_access_token,
                     'pytsite.auth@post_access_token')
     http_api.handle('GET', 'auth/access-token/<token>', _http_api.get_access_token,
                     'pytsite.auth@get_access_token')
     http_api.handle('DELETE', 'auth/access-token/<token>', _http_api.delete_access_token,
                     'pytsite.auth@delete_access_token')
 
-    # User HTTP API handlers
+    # User HTTP API
     http_api.handle('GET', 'auth/user/<uid>', _http_api.get_user, 'pytsite.auth@get_user')
     http_api.handle('PATCH', 'auth/user/<uid>', _http_api.patch_user, 'pytsite.auth@patch_user')
 
-    # Following HTTP API handlers
+    # Following HTTP API
     http_api.handle('POST', 'auth/follow/<uid>', _http_api.follow_user, 'pytsite.auth@follow_user')
-    http_api.handle('DELETE', 'auth/unfollow/<uid>', _http_api.unfollow_user, 'pytsite.auth@unfollow_user')
+    http_api.handle('DELETE', 'auth/follow/<uid>', _http_api.unfollow_user, 'pytsite.auth@unfollow_user')
 
-    # Common routes
+    # Routes
     bp = base_path()
     router.add_rule(bp + '/sign-in/<driver>', 'pytsite.auth@sign_in')
     router.add_rule(bp + '/sign-in/<driver>/post', 'pytsite.auth@sign_in_submit', methods='POST')
@@ -63,6 +63,7 @@ def __init():
     events.listen('pytsite.setup', _eh.setup)
     router.on_dispatch(_eh.router_dispatch, -999)
     router.on_response(_eh.router_response)
+    http_api.on_pre_request(_eh.http_api_pre_request)
 
     # Console commands
     console.register_command(AuthConsoleCommand())