+++ pytsite/odm_http_api/__init__.py
@@ -1,4 +1,4 @@
-"""PytSite ODM HTTP API.
+"""PytSite ODM HTTP API
 """
 
 __author__ = 'Alexander Shepetko'
@@ -10,10 +10,10 @@ def _init():
     from pytsite import http_api
     from . import _http_api
 
-    http_api.handle('POST', 'odm/<model>', _http_api.post_entity, 'pytsite.odm@post_entity')
-    http_api.handle('GET', 'odm/<model>/<uid>', _http_api.get_entity, 'pytsite.odm@get_entity')
-    http_api.handle('PATCH', 'odm/<model>/<uid>', _http_api.patch_entity, 'pytsite.odm@patch_entity')
-    http_api.handle('DELETE', 'odm/<model>/<uid>', _http_api.delete_entity, 'pytsite.odm@delete_entity')
+    http_api.handle('POST', 'odm/entity/<model>', _http_api.post_entity, 'pytsite.odm_http_api@post_entity')
+    http_api.handle('GET', 'odm/entity/<model>/<uid>', _http_api.get_entity, 'pytsite.odm_http_api@get_entity')
+    http_api.handle('PATCH', 'odm/entity/<model>/<uid>', _http_api.patch_entity, 'pytsite.odm_http_api@patch_entity')
+    http_api.handle('DELETE', 'odm/entity/<model>/<uid>', _http_api.delete_entity, 'pytsite.odm_http_api@delete_entity')
 
 
 _init()